requirements.txt
email_validator==2.1.1
fastapi==0.111.0
PyJWT==2.8.0
python-decouple==3.8
uvicorn==0.29.0
sqlmodel==0.0.19
passlib==1.7.4
bcrypt==4.1.3